requirements.txt
crawl4ai>=0.3.0
playwright>=1.40.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
groq>=0.4.0
gunicorn>=21.2.0

